[poc] Switch the implementation of Sink operator to V2.
Sink will now consist of two operators: WriterOperator and CommitterOperator. They are chained by default in streaming mode and behave like Sink V1 in respect to sharing committable instances.
State handling of the writer is similar to V1, however, more logic is being placed inside the SinkWriterStateHandler.
Committer state is being entirely outsourced to the previously introduced CommittableCollector.

The whole translation process is moved inside DataStreamSink by registering a translation listener that expands the single sink to a topology. The SinkTransformation becomes a placeholder for all user configuration of the sink. The transformation is used during expansion to adjust the settings of the transformations of the subtopology.